Use uintptr_t where we can
A lot of code today converts pointers to integers using direct
uint32_t casts. This works fine on 32bit platforms, but makes
the compiler terribly unhappy on 64bit ones.

In practice, it wouldn't matter, since all pointers on the linear
address space on the RPis are 32bit anyway, but let's be nice and
convert them to uintptr_t's instead.

This patch should result in no binary change on 32bit targets.

+++ atag.c
@@ -27,7 +27,7 @@
 
 extern const char *atag_cmd_line;
 extern int conf_source;
-extern uint32_t _atags;
+extern uintptr_t _atags;
 
 int check_atags(const void *atags)
 {
@@ -37,7 +37,8 @@ int check_atags(const void *atags)
 	return -1;
 }
 
-void parse_atags(uint32_t atags, void (*mem_cb)(uint32_t addr, uint32_t len))
+void parse_atags(uintptr_t atags,
+                 void (*mem_cb)(uint32_t addr, uint32_t len))
 {
 	if(atags == 0)
 		return;
