+++ mmio.h
@@ -24,8 +24,8 @@
 
 #include <stdint.h>
 
-void mmio_write(uint32_t reg, uint32_t data);
-uint32_t mmio_read(uint32_t reg);
+void mmio_write(uintptr_t reg, uint32_t data);
+uint32_t mmio_read(uintptr_t reg);
 
 #endif // !MMIO_H
